feat(router): drop `router whoami` and report the allowlist in `status`
Router does not keep a role of its own: it reads the one the Olares edge
injects, so `olares-cli profile whoami` already answers who the CLI is and
`router status` already reports what Router made of that alongside the two
other things a failing verb needs. A second verb for the same record was one
more place to look.

The model allowlist was the one thing only `whoami` printed, so `status` now
prints it when it is non-empty — the difference between a model that is
missing and a model this person may not call.
