+++ cli/skills/olares-router/references/olares-router-architecture.md
@@ -48,7 +48,7 @@ Two planes, two credentials, and they are not interchangeable.
 Router is installed as an admin-only application. Two consequences:
 
 - A non-admin profile cannot see Router's entrance at all, so every verb reports that Router is not installed. `router status` is the check that distinguishes "not installed" from "not visible to you".
-- Within Router, `whoami` reports the role. Most of the management plane requires `admin`, and not only for writes: providers, the vendor catalog's models, market installs, quotas, users, audit and the usage retention window all refuse a non-admin read. What is left to a non-admin is the model list, the routes and default categories, what is installed, their own keys, their own usage, and calling.
+- Within Router, the role is the Olares role: Router reads it from the identity the edge injects rather than keeping one of its own, so `olares-cli profile whoami` and the identity block of `router status` agree by construction. Most of the management plane requires `admin`, and not only for writes: providers, the vendor catalog's models, market installs, quotas, users, audit and the usage retention window all refuse a non-admin read. What is left to a non-admin is the model list, the routes and default categories, what is installed, their own keys, their own usage, and calling.
 
 A Router user is not an Olares user: the row appears the first time that person's identity reaches Router. `router user list` shows the ones it knows, which is why a freshly created Olares account is absent until it makes a call.
